+++ db/migrate/20200116172705_add_service_owner_to_organisations.rb
@@ -1,0 +1,5 @@
+class AddServiceOwnerToOrganisations < ActiveRecord::Migration[6.0]
+  def change
+    add_column :organisations, :service_owner, :boolean, default: false
+  end
+end